improved luma table accuracy (1024 levels + 3.4% threshold) + new hybrid mode to use linear / perceptual depending on luma

+++ src/main/java/org/marlin/pipe/BlendingContextIntSRGB.java
@@ -55,7 +55,8 @@ final class BlendingContextIntSRGB extends BlendComposite.BlendingContext {
         final int[] dstBuffer = ((DataBufferInt) dstOut.getDataBuffer()).getData();
 
         final int srcScan = (srcIn != null) ? ((IntegerInterleavedRaster) srcIn).getScanlineStride() : 0;
-        final int dstScan = ((IntegerInterleavedRaster) dstOut).getScanlineStride();
+        // use long to make later math ops in 64b.
+        final long dstScan = ((IntegerInterleavedRaster) dstOut).getScanlineStride();
 
 //            final BlendComposite.Blender blender = _blender;
         int pixel;
